+++ kelvin-api/changelog.rst
@@ -10,6 +10,7 @@ v1.5.1 (2021-10-20)
 * Add Open Policy Agent component to components documentation (Bug 53960)
 * The log output of the Open Policy Agent is now written to ``/var/log/univention/ucsschool-kelvin-rest-api/opa.log`` (Bug 53961).
 * The test suite for the ``ucsschool.lib`` component was improved (Bug 53962).
+* Username generation counter can now be raised above 100 (Bug 53987).
 
 v1.5.0 (2021-09-10)
 ...................